[content-service] Remove docker dependency
by moving archive creation back to ws-daemon
