+++ components/ws-daemon/nsinsider/go.mod
@@ -65,4 +65,4 @@ replace k8s.io/controller-manager => k8s.io/controller-manager v0.20.1 // leeway
 
 replace k8s.io/kubectl => k8s.io/kubectl v0.20.1 // leeway indirect from gitpod-core/components/common-go:lib
 
-replace k8s.io/mount-utils => k8s.io/mount-utils v0.20.1 // leeway indirect from gitpod-core/components/common-go:lib
+replace k8s.io/mount-utils => k8s.io/mount-utils v0.20.1 // leeway indirect from components/common-go:lib
